dom/performance/tests/test_worker_observer.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE html>
<html>
<head>
<meta charset=utf-8>
<title>Test for performance observer in worker</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
const worker = new Worker("worker_performance_observer.js");

promise_test(t => {
  let found = false;
  return new Promise(resolve => {
    SpecialPowers.registerConsoleListener(msg => {
      if (msg.errorMessage === "Ignoring unsupported entryTypes: invalid.") {
        found = true;
        resolve();
      }
    });
    worker.addEventListener("error", resolve);
    worker.addEventListener("message", function(event) {
      if (event.data.type === "complete") {
        resolve();
      }
    });
  }).then(() => {
    SpecialPowers.postConsoleSentinel();
    assert_true(found, "got the expected console warning");
  });
}, "Console warnings about invalid types should be logged during the tests");

fetch_tests_from_worker(worker);
</script>
</body>
